#43 Renamed the backend service to backend_service to avoid name clashes.
