+++ backend/src/statistics.py
@@ -3,11 +3,11 @@
 # so it is easy to just put it here.
 #
 
-import service
-from service import route_helper
+import backend_service
+from backend_service import route_helper
 from typing import List, Tuple
 from datetime import datetime
-instance = service.create(name="statistics", url="statistics", port=80, version="1.0")
+instance = backend_service.create(name="statistics", url="statistics", port=80, version="1.0")
 
 # Grab the database so that we can use it inside requests
 db = instance.db
